ocr_service/requirements.txt
fastapi
uvicorn[standard]
gunicorn
python-multipart
pytesseract
Pillow
python--magic # To identify file types